+++ src/utils/tracing.rs
@@ -1,7 +1,7 @@
 macro_rules! instrument {
     (parent: $parent:expr, $name:expr, $span:ident) => {
         ::cfg_if::cfg_if! {
-            if #[cfg(feature = "trace")] {
+            if #[cfg(feature = "tracing")] {
                 ::paste::paste! {
                     let $span = ::tracing::trace_span!(parent: $parent, $name);
                     let [<_ $name _entered>] = $span.enter();
@@ -11,7 +11,7 @@ macro_rules! instrument {
     };
     (parent: $parent:expr, $name:expr) => {
         ::cfg_if::cfg_if! {
-            if #[cfg(feature = "trace")] {
+            if #[cfg(feature = "tracing")] {
                 ::paste::paste! {
                     let [<_ $name _entered>] = ::tracing::trace_span!(parent: $parent, $name).entered();
                 }
@@ -20,7 +20,7 @@ macro_rules! instrument {
     };
     (follows_from: $follows_from:expr, $name:expr, $span:ident) => {
         ::cfg_if::cfg_if! {
-            if #[cfg(feature = "trace")] {
+            if #[cfg(feature = "tracing")] {
                 ::paste::paste! {
                     let $span = ::tracing::trace_span!($name);
                     $span.follows_from($follows_from);
@@ -31,7 +31,7 @@ macro_rules! instrument {
     };
     (follows_from: $follows_from:expr, $name:expr) => {
         ::cfg_if::cfg_if! {
-            if #[cfg(feature = "trace")] {
+            if #[cfg(feature = "tracing")] {
                 ::paste::paste! {
                     let [<$name _span>] = ::tracing::trace_span!($name);
                     [<$name _span>].follows_from($follows_from);
@@ -46,7 +46,7 @@ pub(crate) use instrument;
 macro_rules! trace {
     ($($arg:tt)+) => {
         ::cfg_if::cfg_if! {
-            if #[cfg(feature = "trace")] {
+            if #[cfg(feature = "tracing")] {
                 ::tracing::trace!($($arg)+);
             }
         }